+++ cogs/legacy/twokindsrsscog.py
